[WEB-107] changed the folder strcture for constants
