+++ src/components/Home/Reviews/Reviews.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { REVIEWS_MAIN_HEADING, REVIEWS_SUB_HEADING } from "@/constants/reviews";
+import { REVIEWS_MAIN_HEADING, REVIEWS_SUB_HEADING } from "@/constants/home";
 import { REVIEWS } from "@/prototypes/reviews";
 import { ReviewCard } from ".";
 import Reveal from "@/components/Animation/Reveal";
