+++ src/components/LayoutComponents/Navbar/MenuBanner.tsx
@@ -1,4 +1,4 @@
-import { BANNER_DETAILS } from "@/constants/navbar";
+import { BANNER_DETAILS } from "@/constants/home";
 import Link from "next/link";
 import React from "react";
 import { LetsTalkButton } from "../../Reusable";
