--- src/constants/awards/index.ts
@@ -1,1 +1,0 @@
-export const AWARDS_HEADING = "Awards & Recognitions";
