--- src/constants/case_studies/index.ts
@@ -1,3 +1,0 @@
-export const CASE_STUDIES_HEADING = "Case studies & stories";
-export const CASE_STUDIES_SUBHEADING =
-  "As a digital commerce development agency, we work together across the globe to make a difference";
