--- src/constants/clients/index.ts
@@ -1,2 +1,0 @@
-export const CLIENT_IMG_WIDTH = 288;
-export const CLIENT_IMG_HEIGHT = 88;
