--- src/constants/commerce/index.ts
@@ -1,4 +1,0 @@
-export const COMMERCE_MAIN_HEADING =
-  "Leverage modern commerce to create incredible consumer experiences";
-export const COMMERCE_SUB_HEADING =
-  "We push ourselves. We push technology. And we push the boundaries of conventional commerce.";
