--- src/constants/consultation/index.ts
@@ -1,2 +1,0 @@
-export const MAIN_HEADING = "How can we help you?";
-export const SUB_HEADING = "Get all digital commerce services under one roof.";
