--- src/constants/copyrights/index.ts
@@ -1,2 +1,0 @@
-export const COPYRIGHTS_INFO =
-  "Magneto IT Solutions © 2024. All Rights Reserved.";
