--- src/constants/digital_platforms/index.ts
@@ -1,3 +1,0 @@
-export const DIGITAL_PLATFORM_HEADING = "Digital Commerce Platform";
-export const DIGITAL_PLATFORM_DESCRIPTION =
-  "We will recommend and propose a digital commerce platform that is most aligned with your business vision and objectives.\n\n When you’ve reached a crossroads and are unsure what to do next, we will use data-informed decisions that turn your transactional experience into a journey driving absolute customer value";
