--- src/constants/get_in_touch/index.ts
@@ -1,3 +1,0 @@
-export const GET_IN_TOUCH_MAIN_HEADING = "GET IN TOUCH";
-export const GET_IN_TOUCH_SUB_HEADING =
-  "We can only comprehend each other if we communicate well. Allow us to get a better understanding of your needs or queries. Present us the chance to assist you.";
