--- src/constants/hero/index.ts
@@ -1,4 +1,0 @@
-export const HERO_MAIN_HEADING = "Empowering Digital Commerce Transformation";
-export const HERO_SUB_HEADING = `Retailers   |   Distributors   |   Manufacturers  `;
-
-export const HERO_BUTTON_TEXT = "GET A FREE QUOTE";
