+++ src/constants/home/index.ts
@@ -1,0 +1,70 @@
+// Navbar
+export const SOLUTIONS = "Solutions";
+export const PLATFORM = "Platform";
+export const RESOURCES = "Resources";
+export const CUSTOMER = "Customer";
+export const INDUSTRY = "Industry";
+
+export const BANNER_DETAILS = {
+  title: "Work with Magneto",
+  description:
+    "Allow us to get a better understanding of your digital commerce needs.",
+};
+
+export const CONTACT_US = "CONTACT US";
+
+// Hero
+export const HERO_MAIN_HEADING = "Empowering Digital Commerce Transformation";
+export const HERO_SUB_HEADING = `Retailers   |   Distributors   |   Manufacturers  `;
+
+export const HERO_BUTTON_TEXT = "GET A FREE QUOTE";
+
+// Commerce
+
+export const COMMERCE_MAIN_HEADING =
+  "Leverage modern commerce to create incredible consumer experiences";
+export const COMMERCE_SUB_HEADING =
+  "We push ourselves. We push technology. And we push the boundaries of conventional commerce.";
+
+// Results Delivered
+
+export const RESULTS_DELIVERED_HEADING = "The results we deliver";
+
+// Case Studies
+
+export const CASE_STUDIES_HEADING = "Case studies & stories";
+export const CASE_STUDIES_SUBHEADING =
+  "As a digital commerce development agency, we work together across the globe to make a difference";
+
+// Digital Platform
+export const DIGITAL_PLATFORM_HEADING = "Digital Commerce Platform";
+export const DIGITAL_PLATFORM_DESCRIPTION =
+  "We will recommend and propose a digital commerce platform that is most aligned with your business vision and objectives.\n\n When you’ve reached a crossroads and are unsure what to do next, we will use data-informed decisions that turn your transactional experience into a journey driving absolute customer value";
+
+//   Awards
+export const AWARDS_HEADING = "Awards & Recognitions";
+
+// Consultation
+export const MAIN_HEADING = "How can we help you?";
+export const SUB_HEADING = "Get all digital commerce services under one roof.";
+
+// Review
+export const REVIEWS_MAIN_HEADING =
+  "We are trusted by people all across the globe";
+export const REVIEWS_SUB_HEADING =
+  "Check out what our customers have to say about our digital commerce management services";
+
+// Why choose us
+export const WHY_CHOOSE_US_HEADING = "Why choose us?";
+export const WHY_CHOOSE_US_DESCRIPTION =
+  " A team of digital commerce experts working with manufacturers,wholesalers, and retailers globally, turning ideas into unique customer & content-centric experiences using a platform-agnostic approach. \n\n We Push Ourselves. We Push Technology & We Push the Boundaries of Conventional Commerce.";
+
+//   Get in touch
+
+export const GET_IN_TOUCH_MAIN_HEADING = "GET IN TOUCH";
+export const GET_IN_TOUCH_SUB_HEADING =
+  "We can only comprehend each other if we communicate well. Allow us to get a better understanding of your needs or queries. Present us the chance to assist you.";
+
+// CopyRights
+export const COPYRIGHTS_INFO =
+  "Magneto IT Solutions © 2024. All Rights Reserved.";
