--- src/constants/navbar/index.ts
@@ -1,12 +1,0 @@
-export const SOLUTIONS = "Solutions";
-export const PLATFORM = "Platform";
-export const RESOURCES = "Resources";
-export const CUSTOMER = "Customer";
-export const INDUSTRY = "Industry";
-export const BANNER_DETAILS = {
-  title: "Work with Magneto",
-  description:
-    "Allow us to get a better understanding of your digital commerce needs.",
-};
-
-export const CONTACT_US = "CONTACT US";
