--- src/constants/results/index.ts
@@ -1,1 +1,0 @@
-export const RESULTS_DELIVERED_HEADING = "The results we deliver";
