--- src/constants/reviews/index.ts
@@ -1,4 +1,0 @@
-export const REVIEWS_MAIN_HEADING =
-  "We are trusted by people all across the globe";
-export const REVIEWS_SUB_HEADING =
-  "Check out what our customers have to say about our digital commerce management services";
