--- src/constants/why_choose_us/index.ts
@@ -1,3 +1,0 @@
-export const WHY_CHOOSE_US_HEADING = "Why choose us?";
-export const WHY_CHOOSE_US_DESCRIPTION =
-  " A team of digital commerce experts working with manufacturers,wholesalers, and retailers globally, turning ideas into unique customer & content-centric experiences using a platform-agnostic approach. \n\n We Push Ourselves. We Push Technology & We Push the Boundaries of Conventional Commerce.";
